requirements.txt
ics
python-dateutil
requests
google-api-python-client
google-auth-oauthlib
django-environ